+++ packages/frontend-2/components/billing/Alert.vue
@@ -32,6 +32,7 @@ import dayjs from 'dayjs'
 import { graphql } from '~/lib/common/generated/gql'
 import {
   type BillingAlert_WorkspaceFragment,
+  type WorkspaceHeader_WorkspaceFragment,
   WorkspacePlanStatuses
 } from '~/lib/common/generated/gql/graphql'
 import { useBillingActions } from '~/lib/billing/composables/actions'
@@ -53,7 +54,7 @@ graphql(`
 `)
 
 const props = defineProps<{
-  workspace: BillingAlert_WorkspaceFragment
+  workspace: BillingAlert_WorkspaceFragment | WorkspaceHeader_WorkspaceFragment
   actions?: Array<AlertAction>
   condensed?: boolean
 }>()
