+++ packages/frontend-2/layouts/default.vue
@@ -8,7 +8,7 @@
       <div class="relative flex h-[calc(100dvh-3rem)]">
         <DashboardSidebar />
 
-        <main class="w-full h-full overflow-y-auto simple-scrollbar pt-8 pb-16">
+        <main class="w-full h-full overflow-y-auto simple-scrollbar pt-4 lg:pt-8 pb-16">
           <div class="container mx-auto px-6 md:px-8">
             <slot />
           </div>
